Extract flights-contracts and catalog booking-engine contracts
Close the remaining gaps in the *-contracts pattern (ADR-0002 / ADR-0003).

flights was the one module with a contract surface and no contracts package.
Add @voyantjs/flights-contracts (zod-only) owning the pure flight SourceAdapter
contract, request/response schemas, post-book types, and reference-data shapes.

Move catalog's pure booking-engine contracts (booking-engine/contracts — the
BookingDraft + V1 engine schemas — and booking-engine/promotions-contract) into
@voyantjs/catalog-contracts; they were previously trapped in the catalog runtime.

Both runtime packages re-export from the contract packages, so existing
@voyantjs/flights/{contract,reference}/* and @voyantjs/catalog/booking-engine/*
import paths are unchanged. Moved files are byte-identical relocations.

flights' snapshot.ts stays in the runtime: it depends on catalog's
CaptureSnapshotInput/PricingBasis, which live in catalog runtime files. Carving
those pure shapes into catalog-contracts is a tracked follow-up.
