+++ .changeset/flights-and-booking-engine-contracts.md
@@ -1,0 +1,31 @@
+---
+"@voyantjs/flights-contracts": minor
+"@voyantjs/catalog-contracts": minor
+"@voyantjs/flights": minor
+"@voyantjs/catalog": minor
+---
+
+Extract two more contract surfaces into lightweight packages, closing the
+remaining gaps in the `*-contracts` pattern (ADR-0002 / ADR-0003).
+
+`@voyantjs/flights-contracts` (new, zod-only) now owns the pure flight
+`SourceAdapter` contract, request/response schemas, post-book types, and the
+reference-data shapes (`contract/{types,adapter,schemas,post-book-types}`,
+`reference/{contract,static-bundle}`), so flight-provider adapter authors and
+external consumers can integrate without the flights runtime (Drizzle/DB).
+
+`@voyantjs/catalog-contracts` gains the pure booking-engine contracts —
+`booking-engine/contracts` (the `BookingDraft` + V1 engine schemas) and
+`booking-engine/promotions-contract` — which were previously trapped in the
+catalog runtime.
+
+The runtime `@voyantjs/flights` and `@voyantjs/catalog` packages re-export from
+the contract packages, so existing `@voyantjs/flights/contract/*`,
+`@voyantjs/flights/reference/*`, and `@voyantjs/catalog/booking-engine/*` import
+paths are unchanged.
+
+Note: `@voyantjs/flights`' `snapshot.ts` stays in the runtime for now — it
+depends on catalog's `CaptureSnapshotInput` / `PricingBasis`, which still live in
+catalog runtime files (`services/snapshot-service.ts`, `snapshot/schema.ts`).
+Carving those pure shapes into `catalog-contracts` (which would let the flight
+snapshot move too) is a tracked follow-up.
