+++ packages/flights-contracts/src/index.ts
@@ -1,0 +1,6 @@
+export * from "./contract/adapter.js"
+export * from "./contract/post-book-types.js"
+export * from "./contract/schemas.js"
+export * from "./contract/types.js"
+export * from "./reference/contract.js"
+export * from "./reference/static-bundle.js"
